listed state and refactorred app to new component
