+++ 10-ReactOptimisation/src/components/Counter/Configurator.jsx
@@ -1,0 +1,25 @@
+import { useState } from "react";
+import { log } from "../../log";
+
+export default function Configurator({ onSet }) {
+  log("<Configurator /> rendered", 1);
+
+  const [enteredNumber, setEnteredNumber] = useState(0);
+
+  function handleChange(event) {
+    setEnteredNumber(+event.target.value);
+  }
+
+  function handleSetClick() {
+    onSet(enteredNumber);
+    setEnteredNumber(0);
+  }
+
+  return (
+    <section id="configure-counter">
+      <h2>Set Counter</h2>
+      <input type="number" onChange={handleChange} value={enteredNumber} />
+      <button onClick={handleSetClick}>Set</button>
+    </section>
+  );
+}
