+++ codeTestSite.js
@@ -1,0 +1,1 @@
+// this file is added as a play test area for JS code functionality
